Add express and aws-serverless-express to app.
