+++ search/docs/STAC-standalone.yaml
