+++ search/docs/WFS3core+STAC.yaml
