--- search/lib/index.js
@@ -1,2 +1,0 @@
-
-module.export = {};
