+++ search/package.json
@@ -8,8 +8,10 @@
   "license": "Apache-2.0",
   "dependencies": {
     "ajv": "^6.5.1",
+    "aws-serverless-express": "^3.3.6",
     "axios": "^0.18.1",
     "build-url": "^1.1.2",
+    "express": "^4.17.1",
     "js-yaml": "^3.13.1",
     "lodash": "^4.17.15",
     "mime-types": "^2.1.19",
